Clarify that the password is unique
"The password" and "the name" makes it sound like something which exists. There has been confusing by users who have attempted to use their Obsidian or forum accounts.

This new wording is intended to make it clear that this password is unique to your Synced vault. And that the Vault name is something the user chooses as well.